Linter and Test issue fixes
